Don't bail out from gen-manifests early
make gen-manifests would bail out without setting the image when it was
set to false.
To fix it that section is in its own condition, and the rest of the
script can continue if need. Also the default is now true.

Another addition to the Makefile is that we now don't need to specify
the WORKFLOW_ID argument but instead can call it like a target:

   make move2kube gen-manifests

Signed-off-by: Roy Golan <rgolan@redhat.com>