Pass the full github sha to the make targets in github actions
The Makefile default is to use a short 8 chars of the revision, while
the github action is relying on the full github version using the {{
github.sha }} context variable, which is available every where.
We could potentially change the short sha in the makefile, but it is
convenient for developers, hence will leave this untouched for now.

Signed-off-by: Roy Golan <rgolan@redhat.com>