Rename generated secret prefix to 00-
In order to differentiate between resources generated by `kn-workflow
gen-manifests` to manifests added manually by the CI, the prefix of the
secret is changed to `00-*`. This will allow identifying easily the
resources not created by sonataflow and also maintain a proper order of
the files.

p.s.
We are no longer using notifications faker service, hence it can be
removed.

Signed-off-by: Moti Asayag <masayag@redhat.com>

rh-pre-commit.version: 2.3.0
rh-pre-commit.check-secrets: ENABLED