Make the m2k build in a Dockerfile so it is built in a container
Prior to this the build was 2 steps where the 1st one was outside of the
container. With this the build is just a step in the container build
phase.

Also, the openapi file of move2kube api is now stable and on disk,
instead of downloading it during the build. We effectively maintain a
version now in git. It means that we control the changes to the api when
needed and are not effected if there are changes upstream.

Signed-off-by: Roy Golan <rgolan@redhat.com>